KDE5FilePicker: Drop 'm_xListener' and related methods
Drop the 'm_xListener' member and methods using it.
The inherited Qt5FilePicker methods do the same thing.

This also avoids that two methods are called when
a file or filter is selected, since the
'QFileDialog::filterSelected' and 'QFileDialog::fileSelected'
signals had been connected to slots in both, the
qt5 and kde5 fpicker constructors.

Change-Id: I60e3eda29e6497f51165e87d0c4ab2e17f209a1c