Make 'setDefaultName()' work again for kde5 fpicker
Add an option to Qt5FilePicker constructor to
say whether the QFileDialog should be a native one or not,
since 'QFileDialog::selectFile' does not preselect the correct
name in the native dialog any more if the
'QFileDialog::DontUseNativeDialog' option has ever been set, i.e.

    QFileDialog fileDialog;
    fileDialog.setOption(QFileDialog::DontUseNativeDialog);
    fileDialog.setOption(QFileDialog::DontUseNativeDialog, false);
    fileDialog.selectFile("test.txt");

will not properly set the name in the native file dialog, which
broke 'setDefaultName' for the KDE5FilePicker.

This makes it work again, even though I think that the underlying
issue is a Qt bug
(s. https://bugreports.qt.io/browse/QTBUG-73682 ).

Change-Id: I99a1e7c97d594925d600fa8eaf3303f9013551c2